fix bugs

+++ src/controller/master.controller.ts
@@ -8,17 +8,18 @@ import {
     MasterId, UpdateMasterBody
 } from "../interfaces/RequestInterfaces";
 import {NextFunction, Response} from "express";
-import {MasterModel} from "../models/master.model";
+import {MasterModel, MasterStatic} from "../models/master.model";
 import {CityModel} from "../models/city.model";
 import {MasterBusyDateModel} from "../models/masterBusyDate.model";
 import Sequelize, {Attributes, FindAndCountOptions} from "sequelize";
+import {Master, MasterCity, City, MasterBusyDate, ROLE, dbConfig} from '../models';
+import uuid from 'uuid';
+import bcrypt from 'bcrypt';
+import mail from "../services/mailServiсe";
+import tokenService from '../services/tokenServiсe';
+import {MasterCityModel} from "../models/masterCity.model";
 
-const {Master, MasterCity, City, MasterBusyDate, ROLE, dbConfig} = require('../models');
 const ApiError = require('../exeptions/api-error')
-const uuid = require('uuid')
-const bcrypt = require('bcrypt')
-const mail = require("../services/mailServiсe");
-const tokenService = require('../services/tokenServiсe')
 const Op = require('Sequelize').Op;
 
 class MasterController {
@@ -32,7 +33,7 @@ class MasterController {
                 param: "citiesId",
                 location: "body"
             }))
-            const isEmailUniq: MasterModel = await Master.findOne({where: {email}})
+            const isEmailUniq: MasterModel | null = await Master.findOne({where: {email}})
             if (isEmailUniq) return next(ApiError.ExpectationFailed({
                 value: email,
                 msg: `Master with this email is already registered`,
@@ -43,7 +44,8 @@ class MasterController {
             const password: string = randomString.slice(0, 6)
             const hashPassword: string = await bcrypt.hash(password, 5)
             const activationLink: string = uuid.v4();
-            const newMaster: MasterModel = await Master.create({
+            // @ts-ignore
+            const newMaster: MasterModel  = await Master.create({
                 name,
                 email,
                 password: hashPassword,
@@ -57,7 +59,7 @@ class MasterController {
                 newMaster.role,
                 password.slice(0, 6)
             )
-            const findOneCity = (cityId: number): Promise<CityModel> => {
+            const findOneCity = (cityId: number): Promise<CityModel | null> => {
                 return new Promise(function (resolve, reject) {
                     resolve(City.findOne({where: {id: cityId}}))
                     reject(ApiError.BadRequest(`city with this id: ${cityId} is not found`))
@@ -68,14 +70,15 @@ class MasterController {
                 .then(results => {
                         results.map(city => {
                             count++
-                            MasterCity.create({masterId: newMaster.id, cityId: city.id})
+                            // @ts-ignore
+                            city && MasterCity.create({masterId: newMaster.id, cityId: city.id})
                                 .then(() => {
                                     if (count === citiesID.length) {
                                         Master.findOne({
                                             where: {email},
                                             attributes: {exclude: ['password', 'activationLink']},
                                             include: [{model: City}]
-                                        }).then((master: MasterModel) => res.status(201).json(master))
+                                        }).then((master: MasterModel | null) => res.status(201).json(master))
                                     }
                                 })
                         })
@@ -92,7 +95,7 @@ class MasterController {
         try {
             const {limit, offset, sortBy, select, cities, filter} = req.query
             const citiesID: "" | string[] | undefined = cities && cities.split(',');
-            const options: Omit<FindAndCountOptions<Attributes<typeof Master>>, "group"> = {}
+            const options: Omit<FindAndCountOptions<Attributes<MasterModel>>, "group"> = {}
             options.where = {}
             if ((filter !== '') && (filter != undefined) && filter) options.where[Op.or] = [
                 {name: {[Op.iLike]: `%${filter}%`}}, {email: {[Op.iLike]: `%${filter}%`}}]
@@ -115,7 +118,7 @@ class MasterController {
                     required: true
                 }]
             }
-            const masters: MasterModel[] = await Master.findAndCountAll(options)
+            const masters: { rows: MasterModel[]; count: number} = await Master.findAndCountAll(options)
             if (!masters) return next(ApiError.BadRequest("Masters not found"))
             res.status(200).json(masters)
         } catch (e) {
@@ -127,7 +130,7 @@ class MasterController {
     async getOneMaster(req: CustomRequest<null, MasterId, null, null>, res: Response, next: NextFunction) {
         try {
             const masterId = req.params.masterId
-            const master: MasterModel = await Master.findOne({
+            const master: MasterModel | null = await Master.findOne({
                     include: {all: true},
                     where: {id: masterId},
                     attributes: {exclude: ['password', 'activationLink']},
@@ -145,7 +148,7 @@ class MasterController {
         try {
             const {id, name, email, citiesId} = req.body
             console.log(id, name, email, citiesId)
-            const isEmailUniq: MasterModel = await Master.findOne({where: {email}})
+            const isEmailUniq: MasterModel | null = await Master.findOne({where: {email}})
             if (isEmailUniq && isEmailUniq.id !== id) return next(ApiError.ExpectationFailed({
                 value: email,
                 msg: `Master with this email is already registered`,
@@ -163,9 +166,10 @@ class MasterController {
                 location: "body"
             }))
             await MasterCity.destroy({where: {masterId: id}})
-            const createMasterCity = (cityId: number): Promise<CityModel> => {
+            const createMasterCity = (cityId: number): Promise<MasterCityModel> => {
                 return new Promise(function (resolve, reject) {
-                    resolve(MasterCity.create({masterId: id, cityId: Number(cityId)}))
+                    // @ts-ignore
+                    resolve(MasterCity.create({masterId: id, cityId: Number(cityId),createdAt: new Date(Date.now()), updatedAt: new Date(Date.now())}))
                     reject(ApiError.BadRequest(`city with this id: ${cityId} is not found`))
                 })
             }
@@ -179,8 +183,8 @@ class MasterController {
                                     include: [{model: City}],
                                     attributes: {exclude: ['password', 'activationLink']}
                                 }
-                            ).then((master: MasterModel) => {
-                                master.update({name, email})
+                            ).then((master: MasterModel | null) => {
+                                master && master.update({name, email})
                                     .then((master: MasterModel) => res.status(200).json(master))
                             })
                         })
@@ -197,11 +201,9 @@ class MasterController {
         try {
             const {masterId} = req.params
             if (!masterId) next(ApiError.BadRequest("id is not defined"))
-            const candidate: MasterModel = await Master.findOne({where: {id: masterId}, include: [{ all: true }]})
+            const candidate: MasterModel | null = await Master.findOne({where: {id: masterId}, include: [{ all: true }]})
             if (!candidate) next(ApiError.BadRequest(`master with id:${masterId} is not defined`))
-/*            const masterBusyDate: MasterBusyDateModel = await MasterBusyDate.destroy({where: {masterId}})
-            const masterCity: MasterCityModel =await MasterCity.destroy({where: {masterId}})*/
-            await candidate.destroy({ force: true })
+            if (candidate) await candidate.destroy({ force: true })
             res.status(200).json({message: `master with id:${masterId} was deleted`, master: candidate})
         } catch (e) {
             console.log(e)
@@ -213,12 +215,14 @@ class MasterController {
         try {
             const {masterId} = req.params
             if (!masterId) next(ApiError.BadRequest("id is not defined"))
-            const master: MasterModel = await Master.findOne({where: {id: masterId}})
+            const master: MasterModel | null = await Master.findOne({where: {id: masterId}})
             if (!master) next(ApiError.BadRequest(`master with id:${masterId} is not defined`))
-            const isApprove: boolean = master.isApproved
-            await master.update({isApproved: !isApprove})
-            await mail.sendApproveMail(master.email, master.isApproved)
-            res.status(200).json({message: `master with id:${masterId} changed status approve`, master})
+            if (master) {
+                const isApprove: boolean = master.isApproved
+                await master.update({isApproved: !isApprove})
+                await mail.sendApproveMail(master.email, master.isApproved)
+                res.status(200).json({message: `master with id:${masterId} changed status approve`, master})
+            }
         } catch (e) {
             next(ApiError.BadRequest(e))
         }
@@ -265,8 +269,7 @@ class MasterController {
                             masterId: master.id,
                             dateTime: String(time.toISOString())
                         }
-                    }).then((busy: MasterBusyDateModel) => {
-                        console.log(!!busy)
+                    }).then((busy: MasterBusyDateModel | null) => {
                         if (busy) resolve(true)
                         resolve(false)
                     })
@@ -304,7 +307,7 @@ class MasterController {
             param: "citiesId",
             location: "body"
         }))
-        const isEmailUniq: MasterModel = await Master.findOne({where: {email}})
+        const isEmailUniq: MasterModel | null = await Master.findOne({where: {email}})
         if (isEmailUniq) return next(ApiError.ExpectationFailed({
             value: email,
             msg: "Master with this email is already registered",
@@ -317,9 +320,9 @@ class MasterController {
             let result: MasterModel = await dbConfig.transaction(async (t: Sequelize.Transaction) => {
                 const findCity = (cityId: number): Promise<CityModel> => {
                     return new Promise((resolve, reject) => {
-                        City.findOne({where: {id: cityId}, transaction: t}).then((city: CityModel) => {
+                        City.findOne({where: {id: cityId}, transaction: t}).then((city: CityModel | null) => {
                                 if (city === null) reject(`City with id: ${cityId} is not found`)
-                                return resolve(city)
+                                else return resolve(city)
                             }
                         ).catch((err: Error) => {
                             console.log(99)
@@ -329,6 +332,7 @@ class MasterController {
                 }
                 let count = 0
                 return new Promise((resolve, reject) => {
+                    // @ts-ignore
                     Master.create({
                         name,
                         email,
@@ -341,7 +345,7 @@ class MasterController {
                                 Promise.all(citiesId.map(cityId => findCity(cityId)))
                                     .then((cities: CityModel[]) => {
                                             cities.map((city: CityModel) => {
-                                                    MasterCity.create({
+                                                MasterCity.create({
                                                         masterId: newMaster.id,
                                                         cityId: city.id
                                                     }, {transaction: t})
@@ -353,24 +357,20 @@ class MasterController {
                                                                 }
                                                             }
                                                         ).catch((err: Error) => {
-                                                        console.log(88)
                                                         console.log(err)
                                                     })
                                                 }
                                             )
                                         }
                                     ).catch((err: Error) => {
-                                    console.log(77)
                                     console.log(err)
                                 })
                             }
                         ).catch((err: Error) => {
-                        console.log(66)
                         console.log(err)
                     })
                 })
             }).catch((e: Error) => {
-                console.log(55)
                 console.log(e)
             })
             console.log(result)
@@ -380,19 +380,20 @@ class MasterController {
             )
 
             interface MasterModelWithCity extends MasterModel {
-                dataValues: CityModel
+                dataValues?: CityModel
             }
 
-            const master: MasterModelWithCity = await Master.findOne(
+            const master: MasterModelWithCity | null = await Master.findOne(
                 {
                     where: {email: result.email},
                     attributes: {exclude: ['password', 'activationLink']},
                     include: [{model: City}]
                 },
             )
-            console.log(master)
-            const token: string = tokenService.generateJwt(master.id, master.email, master.role)
-            return res.status(201).json({token})
+            if (master) {
+                const token: string = tokenService.generateJwt(master.id, master.email, master.role)
+                return res.status(201).json({token})
+            }
         } catch
             (err) {
             console.log(err)
@@ -403,7 +404,7 @@ class MasterController {
     async changeEmail(req: CustomRequest<ChangeEmailBody, null, null, null>, res: Response, next: NextFunction) {
         try {
             const {password, currentEmail, newEmail} = req.body
-            const master: MasterModel = await Master.findOne({where: {email: currentEmail}})
+            const master: MasterModel | null = await Master.findOne({where: {email: currentEmail}})
             if (!master) return next(ApiError.ExpectationFailed({
                 value: currentEmail,
                 msg: "Master is not found or password is wrong",
